refactor: 예약 api (POST /reservation) 테스트 메소드에 request 값 변경 반영 - LoginStep, ReservationStep 으로 코드 중복 제거
