+++ src/test/java/roomescape/test/LoginAcceptanceTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.equalTo;
